requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
Django==3.0.8
django-user-agents==0.4.0
djangorestframework==3.11.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
markdown2==2.3.9
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
promise==2.3
py==1.9.0
pylint==2.5.3
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
PyYAML==5.3.1
pyyml==0.0.2
Rx==1.6.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
ua-parser==0.10.0
user-agents==2.1
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.1.0
